Add MUVERA, so multi vector retrieval uses the index that is already here
Late interaction retrieval is where quality went. A ColBERT style model gives a
document one vector per token and scores a query against it with Chamfer
similarity, the best match in the document for each query token averaged over
the query. It beats one vector against one vector and it normally costs a whole
second system: an index over every token of every document, so a million
documents at a hundred tokens each is a hundred million vector index, plus a
gather and a scoring pass on top of it.

MUVERA maps a set of token vectors to one fixed length vector whose dot product
with a query's approximates Chamfer, with a bound on the error. So it costs an
encode at write time, the partition index that is already here, and chamfer as
the rerank. There is no second index.

The construction cuts the space into buckets with random hyperplanes. A
document's encoding holds the average of the tokens in each bucket, a query's
holds the sum, and their dot product measures each query token against the
document tokens near it. An empty bucket in a document is filled from the
tokens whose own bucket is nearest in Hamming distance, which is what stops a
short document losing badly, and the whole thing repeats with fresh hyperplanes
because whether a query token and its best match land together is a coin toss.

Two things are different from the paper. There is no final projection of the
whole encoding, because RaBitQ already turns a vector into one bit a dimension
with an error bound a random projection does not have, so projecting first would
throw away accuracy to save space that was going to be saved anyway. And the
encodings come out unit length, so that the index's squared distance orders them
the same way the dot product the approximation is stated in does.

The test corpus took a diagnostic to get right and the reasoning is written into
the file. Unrelated documents and unrelated queries measure nothing: over three
hundred of them the best Chamfer score is 0.331 and the thirtieth best is 0.261,
so which ten are the top ten is decided by noise in the third decimal and no
approximation can recover them. A real query set is not like that, it came from
somewhere, so a query here is six of one document's own tokens with noise on
them and the test is whether that document comes back. It does, first, 95 times
in a hundred.

The encode is a quarter of a million multiply adds and it was written as a
single accumulator dot product, which is the mistake this crate already made
once and fixed in the partition index. Eight running totals took a 16 token
document from 187 to 126 microseconds.

+++ crates/yo-vector/src/muvera.rs
@@ -306,7 +306,7 @@ impl Encoder {
                     // difference whether the scaling happens before the
                     // projection or after.
                     let scale = if document {
-                        1.0 / f64::from(counts[k]) as f32
+                        1.0 / counts[k] as f32
                     } else {
                         1.0
                     };
